Set PTL_SANITIZER_TYPE to existing value, if it exists
Prior to CMake 3.21, CMake's set(... CACHE ...) command could override
existing values of a variable:

- https://gitlab.kitware.com/cmake/cmake/-/issues/22038

This was fixed in 3.21 so that setting the cache value would not override
any existing value:

https://gitlab.kitware.com/cmake/cmake/-/merge_requests/6146

If PTL is used as a subproject, any value for PTL_SANITIZER_TYPE
set by the main project will be overriden to the default "thread".
This can lead to build errors due to more than one, incompatible,
sanitization options being used.

Set PTL_SANITIZER_TYPE to the existing value if it exists, avoiding
override, otherwise defaulting to "thread".