Migration test/example only code to "common" library
PTL's Timer and Backtrace interfaces are only used in example/test code,
with Singleton not used anywhere. Whilst useful, they are not required for
use of PTL's core function of tasking, and consumers of PTL will likely
have their own solutions.

Move this code into the "common" library of examples, building it as a
static lib and link to examples consuming Timer/Backtrace. Remove duplicated
fibonacci function in minimal example so that it only has one definition in
utils.hh